requirements.txt
requests
pyyaml
Flask